G23-117: Moved AlgorithmSchedulePrefRequest from Algorithms folder to DTO folder to maintain consistency in folder structure
