+++ backend/Algorithms/Pgy4RotationScheduleGenerator/Constraints/InpatientConsultInJulyAndJanConstraint.cs
@@ -1,5 +1,6 @@
 
 using MedicalDemo.Enums;
+using MedicalDemo.Models.DTO.Pgy4Scheduling;
 using MedicalDemo.Models.Entities;
 
 namespace MedicalDemo.Algorithms.Pgy4RotationScheduleGenerator.Constraints;
